Update namespace and usage for repositories
The changes made in this commit fall primarily under the adjustment of namespace and usage for repositories. The Domain.Repository namespace has been replaced with Infrastructure.Repository, reflecting a shift in the hierarchical structure of the codebase. Additionally, a using directive has been updated in 'Program.cs' to properly make use of the Infrastructure.Repository namespace.